+++ src/index.ts
@@ -1,3 +1,3 @@
-export { MergeStrategies } from "./animation";
+export { Strategies } from "./types";
 export { createController } from "./controller";
 export { EasingFunctions } from "./easings";